renaming and reorganising files
