+++ frontend/src/pages/AccountSelection.tsx
@@ -4,7 +4,7 @@ import { ActiveTwitterAccountContext, TwitterAccount, TwitterAccounts } from '..
 import { ActiveAccountContext } from '../context/ActiveAccountContext';
 import Button from '../components/ui/Button';
 
-const SelectAccount = () => {
+const AccountSelection = () => {
 	const [twitterAccounts, setTwitterAccounts] = useState<TwitterAccounts>([]);
 	const { activeAccount, setActiveUser } = useContext(ActiveAccountContext) as ActiveTwitterAccountContext;
 
@@ -58,4 +58,4 @@ const SelectAccount = () => {
 	);
 };
 
-export default SelectAccount;
+export default AccountSelection;
