+++ frontend/src/pages/Dashboard/Content.tsx
@@ -1,16 +1,16 @@
-import TweetBox from '../../components/layout/TweetBox';
-import DashboardTweets from './DashboardTweets';
+import TweetBox from '../../components/ui/TweetBox';
+import AllTweets from './tweets/AllTweets';
 
-const DashboardContent = () => {
+const Content = () => {
 	return (
 		<div className="col-span-8">
 			<h2 className="text-xl font-bold mb-8">Dashboard</h2>
 			<TweetBox />
 			<hr className="my-8" />
 			<h2 className="text-xl font-bold mb-4">Tweets</h2>
-			<DashboardTweets />
+			<AllTweets />
 		</div>
 	);
 };
 
-export default DashboardContent;
+export default Content;
